// plugins/docusaurus-plugin-typedoc-api/src/types.ts
export interface ApiOptions {
  banner: string;
  breadcrumbs: boolean;
  gitRefName: string;
  minimal: boolean;
  pluginId: string;
  scopes: string[];
}

export interface TOCItem {
  readonly value: string;
  readonly id: string;
  readonly level: number;
}

// REFLECTIONS

export interface PackageReflectionGroupEntry {
  index: boolean;
  label: string;
  reflection: JSONOutput.ProjectReflection;
  urlSlug: string;
}

export interface PackageReflectionGroup {
  entryPoints: PackageReflectionGroupEntry[];
  packageName: string;
  packageVersion: string;
  changelogPath?: string;
  readmePath?: string;
}